# cars/public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API cars symfony5",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1/",
            "description": "cars API"
        }
    ],
    "paths": {
        "/car/id/{id}": {
            "get": {
                "operationId": "222678f7de9a20a73dc57d2adbb8eee1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of car",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get car by slug",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Car"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/car/slug/{slug}": {
            "get": {
                "operationId": "5036d76cba6ce500e58e45ebe5391732",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "slug of car",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get car by slug",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Car"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cars/enabled": {
            "get": {
                "operationId": "06c8f9170b9d604a5a3e3fd62447e03f",
                "responses": {
                    "200": {
                        "description": "get all cars enabled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Car"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cars": {
            "get": {
                "operationId": "0fee92465870c691a4d9493e0146b91f",
                "responses": {
                    "200": {
                        "description": "get all cars",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Car"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/car/create": {
            "post": {
                "operationId": "fd6d19a56b25ea9ec825e5a517af7d37",
                "requestBody": {
                    "$ref": "#/components/requestBodies/InsertCar"
                },
                "responses": {
                    "200": {
                        "description": "inser car",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "response of insering car",
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/car/update/{id}": {
            "put": {
                "operationId": "b363cec8ba03caaa726d3122105d19c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of car",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateCar"
                },
                "responses": {
                    "200": {
                        "description": "update car",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "response of updating car",
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/car/delete/{id}": {
            "delete": {
                "operationId": "bf1dfba8f932bce11561e50692331fbc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of car",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "update car",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "response of deleting car",
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Car": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "mark": {
                        "type": "string"
                    },
                    "model": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "year": {
                        "type": "integer"
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "nullable": true,
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "requestBodies": {
            "InsertCar": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "mark",
                                "model",
                                "description",
                                "slug",
                                "country",
                                "city",
                                "year",
                                "enabled"
                            ],
                            "properties": {
                                "mark": {
                                    "type": "string"
                                },
                                "model": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "slug": {
                                    "type": "string"
                                },
                                "country": {
                                    "type": "string"
                                },
                                "city": {
                                    "type": "string"
                                },
                                "year": {
                                    "type": "integer"
                                },
                                "enabled": {
                                    "type": "boolean"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "UpdateCar": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "mark": {
                                    "type": "string"
                                },
                                "model": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "slug": {
                                    "type": "string"
                                },
                                "country": {
                                    "type": "string"
                                },
                                "city": {
                                    "type": "string"
                                },
                                "year": {
                                    "type": "integer"
                                },
                                "enabled": {
                                    "type": "boolean"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Class CarController",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}